Moves toast components into the components folder
