+++ App/components/toast/ToastConfig.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Colors } from '../../App/theme'
+import { Colors } from '../../theme'
 
 import BaseToast from './BaseToast'
